# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "\nA clean, modern, and high-performance REST API wrapper over the legacy Urja Meter Ops portal.\n\n### Features\n- **In-Memory Cache & Query Engine**: Instant sub-millisecond filtering, search, and pagination over all 403 meters and 40 transformers.\n- **Geo-Spatial Query Layer**: Locate meters near any geographic coordinates using the Haversine formula (`/api/v1/meters/nearby`).\n- **Network Hierarchy Reconstructor**: Reconstructs the electrical grid topology from Zone down to Meter leaf nodes (`/api/v1/hierarchy`).\n- **Telemetry Data Cleansing**: Normalizes dirty legacy strings, dates, and null markers into typed ISO-8601 timestamps and floats (`/api/v1/meters/{id}/consumption`).\n- **Grid Analytics & Anomalies**: Grid-wide health metrics, manufacturer distribution, and anomaly counts (`/api/v1/analytics/summary`).\n- **Cryptographic Bulk Signer**: Automates HMAC-SHA256 request signing for legacy `/portal/export`.\n- **Interactive Operations Dashboard**: Serves a modern dark-mode web application with an interactive Leaflet map and Chart.js graphs at `/`.\n    ",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login Endpoint",
        "description": "Authenticate specific credentials with the legacy portal.",
        "operationId": "login_endpoint_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "List Meters",
        "description": "Retrieve normalized meters from cache with multi-attribute filtering, search, and pagination.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query for meter ID or serial number",
              "title": "Q"
            },
            "description": "Search query for meter ID or serial number"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by operational status (Active, Decommissioned, Suspended)",
              "title": "Status"
            },
            "description": "Filter by operational status (Active, Decommissioned, Suspended)"
          },
          {
            "name": "make",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by meter manufacturer (e.g. HPL, Genus, Secure)",
              "title": "Make"
            },
            "description": "Filter by meter manufacturer (e.g. HPL, Genus, Secure)"
          },
          {
            "name": "dt_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Distribution Transformer code",
              "title": "Dt Code"
            },
            "description": "Filter by Distribution Transformer code"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Meters Api V1 Meters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/nearby": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get Nearby Meters",
        "description": "Geo-spatial query layer: Locate meters within a specified radius from GPS coordinates.\nUses the great-circle Haversine formula and returns meters sorted by proximity.",
        "operationId": "get_nearby_meters_api_v1_meters_nearby_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90.0,
              "minimum": -90.0,
              "description": "Latitude of the search center coordinate",
              "title": "Lat"
            },
            "description": "Latitude of the search center coordinate"
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180.0,
              "minimum": -180.0,
              "description": "Longitude of the search center coordinate",
              "title": "Lng"
            },
            "description": "Longitude of the search center coordinate"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100.0,
              "exclusiveMinimum": 0.0,
              "description": "Search radius in kilometers",
              "default": 5.0,
              "title": "Radius Km"
            },
            "description": "Search radius in kilometers"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional filter by meter status (e.g. Active)",
              "title": "Status"
            },
            "description": "Optional filter by meter status (e.g. Active)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum number of nearby meters to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of nearby meters to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyMetersListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "tags": [
          "Meters"
        ],
        "summary": "Get Meter Details",
        "description": "Retrieve details for a specific smart meter.",
        "operationId": "get_meter_details_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "AuthLoginRequest"
      },
      "AuthLoginResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "session_token": {
            "type": "string",
            "title": "Session Token"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "session_token",
          "message"
        ],
        "title": "AuthLoginResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}